feat: build account selector UI
The motivation for the account selector is to enable users of the
Hypercerts app to execute certain transactions via a Safe. Currently
only the connected wallet can issue transactions.

The modal will fetch the Safe accounts that the currently connected
wallet is a member of and display them alongside the currently connected
wallet. Any selection made will just visually change the currently
connected wallet, but will have no impact on the behavior of the app
when sending transactions just yet.

Future commits will fully implement this feature.