feat: set name and image url for selected account
This patch adds support for the currently selected account to initiate a
change of user data (display name and image uri).

The message is written out the Safe Transaction Service where users can
facilitate the signature flow. Then the safe message hash is sent to the
Hypercerts API which keeps track of the signature request.

Since this flow is so different to the EOA signing, it can't reuse the
EOA signing flow and at the end of the transaction it can't make the
change to the user data right away, which makes this a 1-step flow.