feat(profile): automatically switch to selected account
Without this patch, selecting any other account other than your
connected EOA using the account selector doesn't do anything on your
profile page. However, we want people to be able to cycle through their
accounts, so whenever another account is selected via the account
selector, we want the profile page to redirect to that account.

On the other hand we don't want the profile to automatically switch when
the user is on a profile page that is not his connected EOA or one of
the multisig they're a signer on. In these cases the selector will do
nothing when the account selector is used.